add two segments intersection tests
